Update manager.yml
The actions used to add reviewers was not working anymore, it was failing with

```
 Error: HttpError: Reviews may only be requested from collaborators. One or more of the teams you specified is not a collaborator of the [organization/repository] repository.
```
even if the collaborator is part of the repo and org.


Other folks were also noticing this error : https://github.com/delivery-much/actions-assigner/issues/7#issuecomment-1439246668 so I don't think it's something on my end. I updated with another one that does the same thing.